pyproject.toml
[tool.poetry]
name = "library"
version = "0.1.0"
description = ""
authors = ["goaho7 <goaho7@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
alembic = "^1.13.3"
pydantic-settings = "^2.6.0"
asyncpg = "^0.30.0"



[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 100
skip-string-normalization = false
target-version = ['py312']